book-app/requirements.txt
Flask==0.12
google-cloud-bigquery
pandas==1.0.5